pyproject.toml
[tool.poetry]
name = "cancer-transcriptome-base"
version = "0.6.0"
description = "A comprehensive cancer transcriptomics platform with working SOTA queries and patient-specific databases"
authors = ["Your Name <toni.wagner@vaudience.ai>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.0"
numpy = "^1.24.0"
psycopg2-binary = "^2.9.5"
gtfparse = "^1.3.0"
sqlalchemy = "^2.0.0"
python-dotenv = "^1.0.0"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
jupyterlab = "^4.0.0"
rich = "^13.0.0"
tqdm = "^4.65.0"
requests = "^2.31.0"
aiohttp = {extras = ["speedups"], version = "^3.9.1"}  # Added with speedups
obonet = "^0.3.0"
networkx = "^3.1.0"
lxml = "^5.3.1"
levenshtein = "^0.27.1"
scipy = "^1.15.3"
httpx = "<0.28"
pyarrow = "^22.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
pytest-asyncio = "^0.23.0"  # Added pytest-asyncio
mypy = "^1.5.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"  # Added asyncio mode configuration
markers = [
  "integration: marks tests that require integration with external services or databases",
  "unit: marks unit tests",
  "slow: marks tests that take longer than 1 second to complete",
  "asyncio: mark test as async/await test"  # Added asyncio marker
]

[tool.pylance]
reportMissingImports = "warning"
reportMissingModuleSource = "warning"

[tool.pyright]
include = ["src", "scripts"]
exclude = ["**/__pycache__"]
reportMissingImports = true
reportMissingModuleSource = true
useLibraryCodeForTypes = true
